Update LICENSE to reflect GGJ licensing
The licensing on GGJ for the game and its source files were submitted under the [Creative Commons Attribution-NonCommercial-ShareAlike v4.0](https://creativecommons.org/licenses/by-nc-sa/4.0/deed.en) license, but the repository was MIT License.
To avoid any potential issues in case people would like to make use of this project, the license for the contents of this repo. has been updated accordingly.